mod: refactor file path to avoid conflict
