+++ src/main/java/Models/nvd2306/Event.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package Models;
+package Models.nvd2306;
 
 import java.sql.Date;
 import java.sql.Timestamp;
